// generated/src/app/api-models/cap-trans-dtl-qlty-incent.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapTransDtlQltyIncent {
  seqCapTransDtlQltyIncent: number;
  seqCapTransDtl: number;
  seqIncentiveRule: number;
  seqQualityPgm: number;
  seqProvContract: number;
  qualityIncentiveAmount: number;
  qualityIncentivePercent: number;
  qualityCapApply: string;
  incentiveRetro: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}